Add avalon:user:admin rake task
When using an omniauth provider such as google, it is helpful to be able
to assign initial administrators to an Avalon application. This commit
adds a rake task to prompt the user for an email address of a user to
assign the administrator role.

Usage:
```
[avalon@lib-avalon-dev current]$ bundle exec rake avalon:user:admin
Assign user as an administrator
Email address for user: mcritchlow@ucsd.edu
Successfully assigned mcritchlow as an administrator
```

If a user does not exist in the system:
```
[avalon@lib-avalon-dev current]$ bundle exec rake avalon:user:admin
Assign user as an administrator
Email address for user: notauser@example.com
User with email address notauser@example.com not found
```